+++ src/main/java/org/terasology/humanoidcharacters/CharacterAppearanceComponent.java
@@ -1,6 +1,6 @@
 // Copyright 2020 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
-package org.terasology.characters.humanoid;
+package org.terasology.humanoidcharacters;
 
 import org.terasology.engine.entitySystem.Component;
 import org.terasology.engine.network.FieldReplicateType;
