+++ src/main/java/org/terasology/humanoidcharacters/VerticalColorArrayMaterialResolver.java
@@ -1,6 +1,6 @@
 // Copyright 2020 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
-package org.terasology.characters.humanoid;
+package org.terasology.humanoidcharacters;
 
 import com.google.common.collect.ImmutableSet;
 import org.terasology.engine.rendering.assets.material.MaterialData;
