+++ src/test/java/org/terasology/humanoidcharacters/CharacterAppearanceClientSystemTest.java
@@ -1,6 +1,6 @@
 // Copyright 2020 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
-package org.terasology.characters.humanoid;
+package org.terasology.humanoidcharacters;
 
 import org.junit.jupiter.api.Test;
 import org.terasology.nui.Color;
